+++ include/SortDataFile.h
@@ -20,8 +20,8 @@
 using namespace std;
 
 int RandomPivot(int first,int last);
-int Partition(vector< pair<int,int> >& A, int f, int l, string option);
-void SortEvent(vector< pair<int,int> >& A, int f, int l, string option);
+int Partition(vector< pair<int,float> >& A, int f, int l, string option);
+void SortEvent(vector< pair<int,float> >& A, int f, int l, string option);
 void SortData(string fName);
 
 #endif // SORTDATAFILE_H
